pyproject.toml
[tool.poetry]
name = "streamlit-postgresql-project"
version = "0.1.0"
description = "A python project that uses streamlit and postgresql local database."
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
# packages = [{include = "postgres_da_ai_agent"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
aiohttp = "3.8.6"
aiosignal = "1.3.1"
annotated-types = "0.6.0"
anyio = "3.7.1"
async-timeout = "4.0.3"
attrs = "23.1.0"
cachetools = "5.3.2"
certifi = "2023.7.22"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
cryptography = "41.0.5"
diskcache = "5.6.3"
distro = "1.8.0"
filelock = "3.13.1"
FLAML = "2.1.1"
frozenlist = "1.4.0"
fsspec = "2023.10.0"
gptcache = "0.1.42"
guidance = "0.0.64"
h11 = "0.14.0"
hiredis = "2.2.3"
httpcore = "1.0.1"
httpx = "0.25.1"
huggingface-hub = "0.17.3"
idna = "3.4"
joblib = "1.3.2"
more-itertools = "9.1.0"
msal = "1.25.0"
multidict = "6.0.4"
nest-asyncio = "1.5.8"
numpy = "1.26.1"
openai = "0.28.1"
packaging = "23.2"
platformdirs = "3.11.0"
psycopg2 = "2.9.9"
pyautogen = "0.1.14"
pycparser = "2.21"
pydantic = "2.0.3"
pydantic-core = "2.3.0"
pygtrie = "2.5.0"
PyJWT = "2.8.0"
pyparsing = "3.1.1"
python-dotenv = "1.0.0"
python-ulid = "1.1.0"
PyYAML = "6.0.1"
redis = "4.6.0"
redis-om = "0.2.1"
regex = "2023.10.3"
requests = "2.31.0"
safetensors = "0.4.0"
scikit-learn = "1.3.2"
scipy = "1.11.3"
sniffio = "1.3.0"
termcolor = "2.3.0"
threadpoolctl = "3.2.0"
tiktoken = "0.5.1"
tokenizers = "0.14.1"
tqdm = "4.66.1"
transformers = "4.35.0"
types-pyOpenSSL = "23.3.0.0"
types-redis = "4.6.0.10"
typing-extensions = "4.8.0"
urllib3 = "2.0.7"
yarl = "1.9.2"
pgvector = "^0.2.3"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
